refacto: split objectsApi into several files
