+++ ui/components/datasetItemWorkspace/src/lib/api/objectsApi/getObjectsToPreAnnotate.ts
@@ -1,0 +1,42 @@
+/*-------------------------------------
+Copyright: CEA-LIST/DIASI/SIALV/LVA
+Author : pixano@cea.fr
+License: CECILL-C
+-------------------------------------*/
+
+import type { Annotation, BBox } from "@pixano/core";
+import { PRE_ANNOTATION } from "../../constants";
+
+export const getObjectsToPreAnnotate = (objects: Annotation[]): Annotation[] =>
+  objects.filter(
+    (object) => object.data.source_ref.name === PRE_ANNOTATION && !object.ui.review_state,
+  );
+
+//Need to check, but it seems this function applies only to BBox
+export const sortAndFilterObjectsToAnnotate = (
+  objects: Annotation[],
+  confidenceFilterValue: number[],
+  currentFrameIndex: number,
+): Annotation[] => {
+  return objects
+    .filter((object) => {
+      if (object.ui.datasetItemType === "image" && object.is_bbox) {
+        const confidence = (object as BBox).data.confidence || 0;
+        return confidence >= confidenceFilterValue[0];
+      }
+      if (
+        object.ui.datasetItemType === "video" &&
+        object.is_bbox &&
+        object.ui.frame_index === currentFrameIndex
+      ) {
+        const confidence = (object as BBox).data.confidence || 0;
+        return confidence >= confidenceFilterValue[0];
+      }
+      return false; // Ignore objects without bboxes
+    })
+    .sort((a, b) => {
+      const firstBoxXPosition = (a as BBox).data.coords[0] || 0;
+      const secondBoxXPosition = (b as BBox).data.coords[0] || 0;
+      return firstBoxXPosition - secondBoxXPosition;
+    });
+};
